requirements.txt
# Telegram Bot
python-telegram-bot>=20.7
python-telegram-bot[job-queue]>=20.7

# Async HTTP requests
aiohttp>=3.9.0
aiofiles>=23.2.0

# Configuration
python-dotenv>=1.0.0

# Audio processing
mutagen>=1.47.0

# Utilities
tqdm>=4.66.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0
